Implement platform-specific extended attribute handling in ShadowFS
- Introduced `file_operations_linux.go` and `file_operations_other.go` to manage extended attributes with Linux-specific and non-Linux stub implementations.
- Refactored existing `GetXattr`, `SetXattr`, and `RemoveXattr` methods to utilize the new platform-specific files, enhancing code organization and maintainability.
- Added `shadow_helpers_linux.go` and `shadow_helpers_other.go` for file copying operations, optimizing for Linux with zero-copy sendfile and providing a fallback for other platforms.
- Updated `xattr` package with platform-specific implementations for getting, setting, and removing extended attributes, ensuring compatibility across different operating systems.
- Improved `Mkdir` method in `shadow_fuse_directory.go` to use the provided mode for directory creation, ensuring correct permissions are set.

+++ fs/utils/permissions.go
@@ -44,7 +44,8 @@ func EnsureDirPermissions(path string) error {
 
 	// Always chmod to ensure kernel sees correct permissions immediately
 	// This is important for FUSE filesystems where kernel might cache wrong attributes
-	if err := syscall.Chmod(path, requiredMode); err != nil {
+	// Cast to uint32 for cross-platform compatibility (Darwin uses uint16 for Stat_t.Mode)
+	if err := syscall.Chmod(path, uint32(requiredMode)); err != nil {
 		return err
 	}
